requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
